+++ src/gql/queries/Asset.php
@@ -37,21 +37,21 @@ class Asset extends Query
                 'args' => AssetArguments::getArguments(),
                 'resolve' => AssetResolver::class . '::resolve',
                 'description' => 'This query is used to query for assets.',
-                'complexity' => GqlHelper::singleQueryComplexity(),
+                'complexity' => GqlHelper::relatedArgumentComplexity()
             ],
             'assetCount' => [
                 'type' => Type::nonNull(Type::int()),
                 'args' => AssetArguments::getArguments(),
                 'resolve' => AssetResolver::class . '::resolveCount',
                 'description' => 'This query is used to return the number of assets.',
-                'complexity' => GqlHelper::singleQueryComplexity(),
+                'complexity' => GqlHelper::relatedArgumentComplexity()
             ],
             'asset' => [
                 'type' => AssetInterface::getType(),
                 'args' => AssetArguments::getArguments(),
                 'resolve' => AssetResolver::class . '::resolveOne',
                 'description' => 'This query is used to query for a single asset.',
-                'complexity' => GqlHelper::singleQueryComplexity(),
+                'complexity' => GqlHelper::relatedArgumentComplexity()
             ],
         ];
     }