+++ src/helpers/Gql.php
@@ -471,6 +471,37 @@ class Gql
     }
 
     /**
+     * Shorthand for returning the complexity function for a field that will add a single query to execution.
+     *
+     * @param int $baseComplexity The base complexity to use. Defaults to a single query.
+     *
+     * @return callable
+     * @since 3.6.7
+     */
+    public static function relatedArgumentComplexity(int $baseComplexity = GqlService::GRAPHQL_COMPLEXITY_QUERY): callable
+    {
+        return static function($childComplexity, $args) use ($baseComplexity) {
+            $complexityScore = $childComplexity + $baseComplexity;
+            $relatedArguments = ['relatedToAssets', 'relatedToEntries', 'relatedToUsers', 'relatedToCategories', 'relatedToTags'];
+
+            foreach ($relatedArguments as $argumentName) {
+                if (!empty($args[$argumentName])) {
+                    $complexityScore += GqlService::GRAPHQL_COMPLEXITY_QUERY * count((array)$args[$argumentName]);
+                }
+            }
+
+            if (!empty($args['relatedTo'])) {
+                $complexityScore += GqlService::GRAPHQL_COMPLEXITY_QUERY;
+            }
+            if (!empty($args['relatedToAll'])) {
+                $complexityScore += GqlService::GRAPHQL_COMPLEXITY_QUERY;
+            }
+
+            return $complexityScore;
+        };
+    }
+
+    /**
      * Shorthand for returning the complexity function for a field that will generate a single query for every iteration.
      *
      * @return callable